+++ packages/react/src/components/dropdownComponents/select/index.ts
@@ -3,4 +3,13 @@ export * from './hooks';
 export { defaultTexts } from './texts';
 export { getElementIds, defaultFilter } from './utils';
 
-// export * from './types';
+export type {
+  SelectProps,
+  SelectCustomTheme,
+  FilterFunction,
+  TextKey as SelectTextKey,
+  Texts as SelectTexts,
+  TextProvider as SelectTextProvider,
+} from './types';
+
+export type { OptionIterator } from './utils';